UPDATED: Worked on segment weight registration max weight and recent weight
-OM-3034

+++ lib/repositories/course_enrollment_repository.dart
@@ -149,7 +149,7 @@ class CourseEnrollmentRepository {
 
     movementsAndWeights.forEach((workoutElement) {
       courseEnrollmentLatestVersion.classes[workoutElement.classIndex].segments[workoutElement.segmentIndex].sections[workoutElement.sectionIndex]
-          .movements[workoutElement.movementIndex].weight = workoutElement.weight;
+          .movements[workoutElement.movementIndex].weight = workoutElement.weight.toDouble();
     });
 
     courseEnrollmentReference.update({
